refactor(main, schemas): UploadIn, ChatIn 클래스를 schemas로 이동
